issue #6

+++ lib/partition/uncoarsening/refinement/node_separators/localized_fm_ns_local_search.cpp
@@ -106,7 +106,7 @@ EdgeWeight localized_fm_ns_local_search::perform_refinement_internal(const Parti
         
         NodeWeight best_separator  = block_weights[2];
         NodeWeight input_separator = block_weights[2];
-        int best_diff              = abs(block_weights[1]-block_weights[0]);
+        int best_diff              = abs((int)block_weights[1]-(int)block_weights[0]);
         int undo_idx = 0;
 
         std::vector< NodeWeight > best_block_weights(3,0);
@@ -160,7 +160,7 @@ EdgeWeight localized_fm_ns_local_search::perform_refinement_internal(const Parti
                         }
                 }
 
-                int cur_diff = abs(block_weights[1]-block_weights[0]);
+                int cur_diff = abs((int)block_weights[1]-(int)block_weights[0]);
                 if( block_weights[2] < best_separator || (block_weights[2] == best_separator && cur_diff < best_diff)  ) {
                         best_separator              = block_weights[2];
                         undo_idx                    = rollback_info.size();
@@ -214,7 +214,7 @@ EdgeWeight localized_fm_ns_local_search::perform_refinement_internal(const Parti
 
         NodeWeight best_separator  = block_weights[2];
         NodeWeight input_separator = block_weights[2];
-        int best_diff              = abs(block_weights[1]-block_weights[0]);
+        int best_diff              = abs((int)block_weights[1]-(int)block_weights[0]);
         int undo_idx = 0;
 
         int steps_till_last_improvement = 0;
@@ -265,7 +265,7 @@ EdgeWeight localized_fm_ns_local_search::perform_refinement_internal(const Parti
                         }
                 }
 
-                int cur_diff = abs(block_weights[1]-block_weights[0]);
+                int cur_diff = abs((int)block_weights[1]-(int)block_weights[0]);
                 if( block_weights[2] < best_separator || (block_weights[2] == best_separator && cur_diff < best_diff)  ) {
                         best_separator = block_weights[2];
                         undo_idx = rollback_info.size();
